07.13 자바스크립트,AJAX 이용 회원게시물 댓글쓰기(작성자만 삭제 가능) 회원게시물 (작성자만 삭제, 수정 가능)

+++ exspring/src/main/java/com/exam/myapp/bbs/BbsServiceImpl.java
@@ -79,15 +79,19 @@ public class BbsServiceImpl implements BbsService {
 
 	//게시글 삭제
 	@Override
-	public int deleteBbs(int bbsNo) {
+	public int deleteBbs(BbsVo vo) {
 		//
-		BbsVo bbsVo = bbsDao.selectBbs(bbsNo);	//삭제할 게시글 정보 조회
+		BbsVo bbsVo = bbsDao.selectBbs(vo.getBbsNo());	//삭제할 게시글 정보 조회
+		if ( ! bbsVo.getBbsWriter().equals(vo.getBbsWriter())) {
+			return 0;
+		}
+			
 		for (AttachVo attVo : bbsVo.getAttachList() ) {	//게시글의 첨부파일을 하나씩 꺼내서
 			new File(uploadPath, attVo.getAttNewName()).delete();	//디스크에서 첨부파일 삭제
 			attachDao.deleteAttach(attVo.getAttNo());	//테이블에서 첨부파일 삭제
 		}
 		
-		return bbsDao.deleteBbs(bbsNo);
+		return bbsDao.deleteBbs(vo.getBbsNo());	//테이블에서 게시글 삭제
 	}
 
 	//게시글 수정
